requirements.txt
SpeechRecognition>=3.10.4
pyttsx3==2.90
apscheduler==3.10.4
requests==2.31.0
wikipedia==1.4.0
python-dotenv==1.0.0
PyAudio==0.2.14
